Separating session-store from session.js
The motive of this commit is to seperate unrelated logic into separate files.

Also, we are trying to tighten the interface between different modules. So, now
a Session doesn't have an options attr that it passes every where. We will pick
out what is required before interacting with any other module.